update to async-nats with jetstream objectstore and new headers api (#120)
* update to async-nats with jetstream objectstore and new headers api


* remove accidental '#'


* update to async-nats 0.21.0


* synk warning - bring tailwind up to date


* (WIP) async-nats 0.21-0.22 changes


* add json tags to go codegen; update dependencies; clippy; bump codegen to 0.5.1


* misc clippy fixes


* default config


* bump to async-nats 0.22.0


* cleanup prometheus conditional, remove unused function


* use logging in tests instead of println


* tinygo merge


* pr feedback


* rustfmt


* update codegen mssrv to 1.60


* misc cleanup; bump msrv to 1.64 due to async-nats requirement


* bump macros msrv to 1.56.1


* update changelog


* change warning message to recommend upgrade to host v0.59.0


* misc


* bump to 0.6.0


* pr feedback; fix span fields

+++ rpc-rs/src/timestamp.rs
@@ -110,7 +110,7 @@ impl From<SystemTime> for Timestamp {
             .expect("system time before Unix epoch");
         Timestamp {
             sec: d.as_secs() as i64,
-            nsec: d.subsec_nanos() as u32,
+            nsec: d.subsec_nanos(),
         }
     }
 }
